chore(prism-agent): refactoring

+++ prism-agent/service/server/src/main/scala/io/iohk/atala/api/http/model/CollectionStats.scala
@@ -1,3 +1,3 @@
 package io.iohk.atala.api.http.model
 
-case class CollectionStats(totalCount: Int, filteredCount: Int)
+case class CollectionStats(totalCount: Long, filteredCount: Long)
